Change the URL for when there is no WebGL
There is an "official" url now. Over time it 
is supposed to be updated to auto redirect
depending on the browser so Firefox would
redirect to a Firefox upgrade page, Safari
to a Safari upgrade page etc..

Also, make o3d-webgl run in IE6/7/8/9
at least to the point that it comes up with
the message to get a browser that supports
WebGL

Also also note: There a 3 or more cases that
need to be covered for WebGL apps

Case #1: canvas doesn't exist at all. (IE6/7/8)

Case #2: canvas does exist but there's no WebGL

Case #3: canvas does exist, WebGL exists and should
   run but can't. (ie, don't tell the user to get new
   browser in this case)

WebGL is being updated to give info so case #3
can be handled so we can update o3d-webgl
when that's available.


TEST=none
BUG=none

Review URL: http://codereview.chromium.org/3204014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57603 0039d316-1c4b-4281-b951-d872f2087c98