Pull seccomp-sandbox in via DEPS rather than using an in-tree copy
This means changes to the sandbox won't have to be committed twice, to 
both trees. 

This is a retry of r57921, which was committed with git-svn and failed 
to remove the "seccomp" directory. This caused problems when trying 
to "svn checkout" to the same location, and the change was reverted. 
This time I will use SVN to commit the change. 

BUG=none 
TEST=smoke test of running chromium with --enable-seccomp-sandbox 
Review URL: http://codereview.chromium.org/3225010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@58184 0039d316-1c4b-4281-b951-d872f2087c98