We used to show error messages as info bars earlier. Based on UX feedback, we now show the messages within the speech input UI bubble. We also let the user retry recognition without returning the error to the web page unless the user clicked the cancel button or closed the bubble.
For this to work, we now keep the request alive in the SpeechInputManager map until it completes successfully or user explicitly cancelled it. We also keep the SpeechInputBubble object alive, even when the actual bubble window is not visible on screen, and construct the window if required whenever we want to show the recording status or error message on screen.

BUG=53598
TEST=unit_tests --gtest_filter=SpeechInputBubbleControllerTest.*

Review URL: http://codereview.chromium.org/3352018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59166 0039d316-1c4b-4281-b951-d872f2087c98