Update omnibox popup when selection changes.
The original motivation for this cl corresponds to bug 47210: If the user changes the selection by arrowing around in the omnibox, we should detect this as accepting the inline autocomplete suggestion.  Otherwise, if the user subsequently presses ctrl (e.g. as part of the keystroke ctrl-d to perform a forward delete), the omnibox contents are reverted as if the user had never arrowed around.

The other motivation is just to match the behavior on Windows, which detects changes to the selection.

BUG=47210
TEST=none

Review URL: http://codereview.chromium.org/3126024

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@59770 0039d316-1c4b-4281-b951-d872f2087c98