Revert 60368 - Client-side phishing detection: grab snapshot of custom sized view.
Client-side phishing detection needs to be able to take a snapshot of a
particular page that has a constant size and aspect ratio (e.g., 1024x768).
This CL adds a helper function (safe_browsing::GrabPhishingThumbnail)
which re-sizes the view a given size, grabs the snapshot, and then re-sizes
the view back to its previous size.  Note: this function can be slow since
it might re-layout the page twice.

This function will only be called if we are very confident that the current page
is a phishing site in which case we are OK with taking a slow snapshot.

Also, this CL adds a test for the OnMsgPaintAtSize method in the RenderWidget
since it does something very similar than the new Thumbnailer.

BUG=None
TEST=GrabPhishingThumbnail

Review URL: http://codereview.chromium.org/3380001

TBR=noelutz@google.com
Review URL: http://codereview.chromium.org/3439026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@60536 0039d316-1c4b-4281-b951-d872f2087c98