Create a new Navigate() function to (eventually) supercede the existing maze of methods that add and navigate tabs in Browser.
This function is intended to replace at least the following:
Browser::OpenURLAtIndex
Browser::AddTabWithURL
Browser::AddNewContents
... as well as the myriad other functions that call into these.

Right now it's just a standalone thing with a set of unit tests.

A TBD for me is to support singleton tabs. I'm going to work on that next.

BUG=none
TEST=unit tests
Review URL: http://codereview.chromium.org/3591011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@62195 0039d316-1c4b-4281-b951-d872f2087c98