Fix a bug where in having ChromeFrame installed on the machine would cause a non chrome frame email site to automatically logout after logging in. ChromeFrame appends the chromeframe string to the user agent to outgoing requests initiated from IE. However this was only done in the protocol sink patch for top level requests. Sub requests would carry the IE user agent which in this case would cause the web server to get confused and drop the connection.
Fix is to resurrent the IHttpNegotiate vtable patch for the protocol patch as well.

Removed the code which read the patch method from the registry and turned on the moniker
or the httpequiv patch. It is unlikely that we will need this in the future. Cleaned
up the http negotiate patch by removing references to the navigation manager and the
ReportProgress patch which is not needed anymore.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=60745

Bug=60745

Review URL: http://codereview.chromium.org/4247001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@64688 0039d316-1c4b-4281-b951-d872f2087c98