Revert 65465 - testserver.py listens on ephemeral ports by default.
Some SSL UI tests timed out again on 10.5/10.6 bots when landing this, despite the changes made. Different tests this time however.

If --port is specified on the command line, testserver.py will listen on that port, otherwise it will listen on an ephemeral port. If --startup_pipe is specified, the port number is written to the pipe as a 2 byte unsigned int in host order.

TestServer spawns testserver.py to listen on an ephemeral port and reads the port value from the pipe. A fixed port can not be specified.

BUG=56814
TEST=try bots pass

Review URL: http://codereview.chromium.org/4136008

TBR=cbentzel@chromium.org
Review URL: http://codereview.chromium.org/4731003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65488 0039d316-1c4b-4281-b951-d872f2087c98